+++ src/session_state.rs
@@ -1,30 +1,35 @@
 /// This module represents the session state of an MQTT communication session.
-use crate::warn;
+use crate::{warn, QoS};
 use embedded_nal::IpAddr;
-use heapless::{String, Vec};
+use heapless::{LinearMap, String, Vec};
 
 use embedded_time::{
     duration::{Extensions, Seconds},
-    Clock, Instant,
+    Instant,
 };
 
 /// The default duration to wait for a ping response from the broker.
 const PING_TIMEOUT: Seconds = Seconds(5);
 
-pub struct SessionState<C: Clock> {
+pub struct SessionState<Clock: embedded_time::Clock, const MSG_SIZE: usize, const MSG_COUNT: usize>
+{
     pub keep_alive_interval: Option<Seconds<u32>>,
-    ping_timeout: Option<Instant<C>>,
-    next_ping: Option<Instant<C>>,
+    ping_timeout: Option<Instant<Clock>>,
+    next_ping: Option<Instant<Clock>>,
     pub broker: IpAddr,
     pub maximum_packet_size: Option<u32>,
     pub client_id: String<64>,
     pub pending_subscriptions: Vec<u16, 32>,
+    pub pending_publish: LinearMap<u16, Vec<u8, MSG_SIZE>, MSG_COUNT>,
+    pub pending_publish_ordering: Vec<u16, MSG_COUNT>,
     packet_id: u16,
     active: bool,
 }
 
-impl<C: Clock> SessionState<C> {
-    pub fn new<'a>(broker: IpAddr, id: String<64>) -> SessionState<C> {
+impl<Clock: embedded_time::Clock, const MSG_SIZE: usize, const MSG_COUNT: usize>
+    SessionState<Clock, MSG_SIZE, MSG_COUNT>
+{
+    pub fn new(broker: IpAddr, id: String<64>) -> SessionState<Clock, MSG_SIZE, MSG_COUNT> {
         SessionState {
             active: false,
             ping_timeout: None,
@@ -34,6 +39,8 @@ impl<C: Clock> SessionState<C> {
             packet_id: 1,
             keep_alive_interval: Some(59.seconds()),
             pending_subscriptions: Vec::new(),
+            pending_publish: LinearMap::new(),
+            pending_publish_ordering: Vec::new(),
             maximum_packet_size: None,
         }
     }
@@ -44,10 +51,59 @@ impl<C: Clock> SessionState<C> {
         self.keep_alive_interval = Some(59.seconds());
         self.maximum_packet_size = None;
         self.pending_subscriptions.clear();
+        self.pending_publish.clear();
+        self.pending_publish_ordering.clear();
+    }
+
+    /// Called when publish with QoS 1 is called so that we can keep track of PUBACK
+    pub fn handle_publish(&mut self, qos: QoS, id: u16, packet: &[u8]) {
+        // This is not called for QoS 0 and QoS 2 is not implemented yet
+        assert_eq!(qos, QoS::AtLeastOnce);
+
+        let mut buf: Vec<u8, MSG_SIZE> = Vec::from_slice(packet).unwrap();
+        // Set DUP = 1 (bit 3). If this packet is ever read it's just because we want to resend it
+        buf[0] |= 1 << 3;
+        // If this fails and the PUBACK will be received and Client (minimq) should disconnect from server with 0x82 Protocol Error
+        // This behaviour pretty much reverts this message to QoS 0 with a restart if the message is actually delivered
+        let _ = self.pending_publish.insert(id, buf);
+        let _ = self.pending_publish_ordering.push(id);
+    }
+
+    /// Delete given pending publish as the server took ownership of it
+    pub fn handle_puback(&mut self, id: u16) {
+        self.pending_publish.remove(&id);
+        let mut found = false;
+        for i in 0..self.pending_publish_ordering.len() {
+            if found {
+                self.pending_publish_ordering[i - 1] = self.pending_publish_ordering[i];
+            } else {
+                if self.pending_publish_ordering[i] == id {
+                    found = true;
+                }
+            }
+        }
+        self.pending_publish_ordering.pop();
+    }
+
+    /// Indicates if publish with QoS 1 is possible.
+    pub fn can_publish(&self, qos: QoS) -> bool {
+        match qos {
+            QoS::AtMostOnce => true,
+            QoS::AtLeastOnce => self.pending_publish.len() < MSG_COUNT,
+            QoS::ExactlyOnce => false,
+        }
+    }
+
+    pub fn pending_messages(&self, qos: QoS) -> usize {
+        match qos {
+            QoS::AtMostOnce => 0,
+            QoS::AtLeastOnce => self.pending_publish.len(),
+            QoS::ExactlyOnce => 0,
+        }
     }
 
     /// Called whenever an active connection has been made with a broker.
-    pub fn register_connection(&mut self, now: Instant<C>) {
+    pub fn register_connection(&mut self, now: Instant<Clock>) {
         self.active = true;
         self.ping_timeout = None;
 
@@ -97,7 +153,7 @@ impl<C: Clock> SessionState<C> {
     /// # Returns
     /// An error if a pending ping is past the deadline. Otherwise, returns a bool indicating
     /// whether or not a ping should be sent.
-    pub fn handle_ping(&mut self, now: Instant<C>) -> Result<bool, ()> {
+    pub fn handle_ping(&mut self, now: Instant<Clock>) -> Result<bool, ()> {
         // First, check if a ping is currently awaiting response.
         if let Some(timeout) = self.ping_timeout {
             if now > timeout {